move *_of into analysis module
