+++ python/test/analysis/test_analysis.py
@@ -1,15 +1,14 @@
 from typing import List
 import opendp.prelude as dp
-from opendp.analysis import Analysis, unit_of, loss_of
 
 dp.enable_features("contrib")
 
 
 def test_analysis_init():
-    analysis = Analysis.sequential_composition(
+    analysis = dp.Analysis.sequential_composition(
         data=[1, 2, 3],
-        privacy_unit=unit_of(contributions=3),
-        privacy_loss=loss_of(epsilon=3.0),
+        privacy_unit=dp.unit_of(contributions=3),
+        privacy_loss=dp.loss_of(epsilon=3.0),
         split_evenly_over=3,
         domain=dp.domain_of(List[int]),
     )
@@ -24,10 +23,10 @@ def test_analysis_init():
 
 
 def test_analysis_zCDP():
-    analysis = Analysis.sequential_composition(
+    analysis = dp.Analysis.sequential_composition(
         data=[1, 2, 3],
-        privacy_unit=unit_of(contributions=1),
-        privacy_loss=loss_of(epsilon=3.0, delta=1e-6),
+        privacy_unit=dp.unit_of(contributions=1),
+        privacy_loss=dp.loss_of(epsilon=3.0, delta=1e-6),
         split_evenly_over=2,
     )
 
@@ -45,10 +44,10 @@ def test_analysis_zCDP():
 
 
 def test_sc_query():
-    analysis = Analysis.sequential_composition(
+    analysis = dp.Analysis.sequential_composition(
         data=[1, 2, 3],
-        privacy_unit=unit_of(contributions=1),
-        privacy_loss=loss_of(epsilon=3.0, delta=1e-6),
+        privacy_unit=dp.unit_of(contributions=1),
+        privacy_loss=dp.loss_of(epsilon=3.0, delta=1e-6),
         split_evenly_over=2,
         domain=dp.vector_domain(dp.atom_domain(T=int)),
     )
@@ -85,14 +84,15 @@ def test_sc_query():
 
 
 def test_distance_of():
-    assert unit_of(contributions=3) == (dp.symmetric_distance(), 3)
-    assert unit_of(l1=2.0) == (dp.l1_distance(T=float), 2.0)
+    assert dp.unit_of(contributions=3) == (dp.symmetric_distance(), 3)
+    assert dp.unit_of(l1=2.0) == (dp.l1_distance(T=float), 2.0)
 
 
 def test_privacy_loss_of():
-    assert loss_of(epsilon=3.0) == (dp.max_divergence(T=float), 3.0)
-    assert loss_of(rho=2.0) == (dp.zero_concentrated_divergence(T=float), 2.0)
-    assert loss_of(epsilon=2.0, delta=1e-6) == (
+    assert dp.loss_of(epsilon=3.0) == (dp.max_divergence(T=float), 3.0)
+    assert dp.loss_of(rho=2.0) == (dp.zero_concentrated_divergence(T=float), 2.0)
+    assert dp.loss_of(epsilon=2.0, delta=1e-6) == (
         dp.fixed_smoothed_max_divergence(T=float),
         (2.0, 1e-6),
     )
+test_privacy_loss_of()
